aiohttp_mongo/requirements.txt
aiohttp==3.1.0
PyYAML==3.12
motor==1.2.1
uvloop==0.9.1
locustio==0.8.1
coverage==4.5.1